chore: get rid of the 'new_epoch_state' module in kernel.
